move random map generator functions to features/map/random-map-gen
