+++ src/features/map/random-map-gen/generate-monster-type.js
