+++ src/features/world/index.js
@@ -7,8 +7,9 @@ import Monsters         from '../monsters';
 import takeMonstersTurn from '../monsters/actions/take-monsters-turn';
 import Player           from '../player';
 import exploreTiles     from '../player/actions/explore-tiles';
-import generateMonsters from '../../modules/generate-monsters';
-import store            from '../../config/store';
+import generateMonsters from '../map/random-map-gen/generate-monsters';
+
+import store from '../../config/store';
 
 import './styles.scss';
